:sparkles: feat(component-asset): 添加 createTestAssetStore 方法，便于下游测试
